requirements.txt
absl-py==0.8.1
astor==0.8.0
cachetools==3.1.1
certifi==2019.11.28
chardet==3.0.4
gast==0.2.2
google-auth==1.7.1
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.25.0
h5py==2.10.0
idna==2.8
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
numpy==1.17.4
oauthlib==3.1.0
opt-einsum==3.1.0
Pillow==6.2.1
protobuf==3.11.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.3.3
six==1.13.0
tensorboard==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
termcolor==1.1.0
tflearn==0.3.2
tqdm==4.40.0
urllib3==1.25.7
Werkzeug==0.16.0
windows-curses==2.1.0
wrapt==1.11.2
